fix(mobile): fix broken state on share-to-bookmark screen
Three bugs caused the share screen to show null/empty state instead of
navigating after save:

1. `resetShareIntent()` was called in `navigateAway()` before navigation.
   This synchronously cleared the share intent provider state, causing
   `sharedUrl` (derived inline from the live hook) to become null and the
   component to re-render with the empty state while still visible on screen.

2. `_layout.tsx` navigated to `/share` without passing the URL as a route
   param (`router.replace('/share')`), so `params.url` was always undefined
   in share.tsx. The URL was only readable from the volatile `shareIntent`
   hook, which could be reset at any moment.

3. `sharedUrl` was computed inline from the live `shareIntent` hook value on
   every render, so any external reset of the intent (resetOnBackground, etc.)
   would immediately null it out mid-session.

Fix: _layout.tsx now extracts the URL from the intent before navigating and
passes it as a query param (`/share?url=...`). share.tsx captures the URL once
on mount via useState lazy initializer, removing the dependency on the live
share intent entirely. `navigateAway()` no longer calls `resetShareIntent()`.
